requirements.txt
gradio
python-dotenv
openai
anthropic
